move struct derive into own module
