+++ derive/src/derive_enum.rs
@@ -1,0 +1,5 @@
+use syn::Ident;
+
+pub struct DerivedEnum<'a> {
+    name: &'a Ident,
+}
